+++ src/blockly/extensions/var_menu.ts
@@ -1,17 +1,14 @@
 import * as Blockly from "blockly/core";
 
+function nameOf([name]: [string, string]) {
+    return [name, name] as [string, string];
+}
+
 export default function (this: Blockly.Block) {
     const input = this.getInput("DUMMY")!;
-    const menu = new Blockly.FieldDropdown(() => {
-        const vars = window.app.current.variables.map<[string, string]>(([e]) => [e, e]);
-
-        if (window.app.current.name !== "Stage") {
-            for (const variable of window.app.globalVariables) {
-                vars.push([variable, variable]);
-            }
-        }
-
-        return vars;
-    });
+    const menu = new Blockly.FieldDropdown(() => [
+        ...window.app.current.variables.map(nameOf),
+        ...window.app.current.name === "Stage" ? [] : window.app.globalVariables.map(nameOf)
+    ]);
     input.appendField(menu, "VAR");
 };